Separate the e2e external net testing to it's own directory
In this way we avoid having to deal around with the skip/focus
flags, and we have a single parameter to differencitate.
